Refactor result classes into dedicated namespace
