--- lib/ai_toolkit/message_result.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "result_item"
-
-module AiToolkit
-  # Represents a normal message from the LLM.
-  class MessageResult < ResultItem
-    attr_reader :role, :content
-
-    # @param role [String]
-    # @param content [String]
-    def initialize(role:, content:)
-      super()
-      @role = role
-      @content = content
-    end
-  end
-end
