--- lib/ai_toolkit/result_item.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module AiToolkit
-  # Base class for all result items returned by a request.
-  class ResultItem
-    # Constant used to mark the base item type.
-    TYPE = :result_item
-  end
-end
