+++ lib/ai_toolkit/results/result_item.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module AiToolkit
+  module Results
+    # Base class for all result items returned by a request.
+    class ResultItem
+      # Constant used to mark the base item type.
+      TYPE = :result_item
+    end
+  end
+end
