--- lib/ai_toolkit/tool_response.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "result_item"
-
-module AiToolkit
-  # Represents a tool response sent back to the LLM.
-  class ToolResponse < ResultItem
-    attr_reader :tool_use_id, :content
-
-    # @param tool_use_id [String]
-    # @param content [String]
-    def initialize(tool_use_id:, content:)
-      super()
-      @tool_use_id = tool_use_id
-      @content = content
-    end
-  end
-end
